WIP - Add RollupConfig.MaxTxnDelay
If a pending transaction was received more than MaxTxnDelay ago then the
minimum commitment and batch sizes are set to 1 in order to allow the
transaction to make it into a batch.

This allows us to safely increase the minimum batch sizes, saving on gas
costs, knowing that it will not have an outiszed impact on users who
make transactions during times of low transaction volume.