Extract SyncedBlock from ChainState record
While syncing, every time we process a new block we update the
SyncedBlock record. When a row is updated in Badger the entire row is
written to disk. Previously, this meant that every time we injested a
new block we wrote a copy of the entire ChainState record to disk. This
caused massive write amplification for deployments with a large genesis.
SyncedBlock is now stored separately, eliminating the worst of the write
amplification.